refactor: hooks for blueprint feature
