+++ apps/web/src/pages/templates/editor/TemplateEditorPage.tsx
@@ -19,8 +19,8 @@ import { TestWorkflowModal } from '../../../components/templates/TestWorkflowMod
 import { SaveChangesModal } from '../../../components/templates/SaveChangesModal';
 import { useDisclosure } from '@mantine/hooks';
 import { ExecutionDetailsModalWrapper } from '../../../components/templates/ExecutionDetailsModalWrapper';
-import { useSearchParams } from '../../../hooks/use-SearchParams';
-import { BluePrintModal } from '../../../components/templates/BlueprintModal';
+import { useSearchParams } from '../../../hooks/useSearchParams';
+import { BlueprintModal } from '../../../components/templates/BlueprintModal';
 
 export enum ActivePageEnum {
   SETTINGS = 'Settings',
@@ -60,8 +60,6 @@ export default function TemplateEditorPage() {
     onTriggerModalDismiss,
   } = useTemplateController(templateId);
   const isCreateTemplatePage = location.pathname === '/templates/create';
-  const [blueprintId, setBluePrintId] = useState<undefined | string>();
-
   const [showModal, confirmNavigation, cancelNavigation] = usePrompt(isDirty);
 
   const [testWorkflowModalOpened, { close: closeTestWorkflowModal, open: openTestWorkflowModal }] = useDisclosure(
@@ -85,11 +83,6 @@ export default function TemplateEditorPage() {
 
     setActivePage(page);
   }, [searchParams.page]);
-
-  useEffect(() => {
-    const id = localStorage.getItem('blueprintId');
-    setBluePrintId(id === null ? undefined : id);
-  }, [localStorage.getItem('blueprintId')]);
 
   const onConfirmSaveChanges = async (data: IForm) => {
     await onSubmit(data);
@@ -206,7 +199,7 @@ export default function TemplateEditorPage() {
         cancelNavigation={cancelNavigation}
         confirmNavigation={confirmNavigation}
       />
-      <BluePrintModal blueprintId={blueprintId} />
+      <BlueprintModal />
     </>
   );
 }
